update to todays working copy - make keys work from plugins
Add some .gitignore files, and get the plugin conf and conf request working
well. Plugins can request a key to watch, and an enum, and then the module
will call the function on the plugin with the enum when the value changes.
This is a little more than O(1) and so scales when lots of plugins watch
the same key. It also removes the processing of strcmp on each key change
event as these are matched internally using a hash table.