change plugin interface to be more efficient
This patch converts the plugin interface to use static dats for interested,
requires, provides, suggests and prevents. It defines a set of macros for
plugin authors to use to make declaring these nice and simple.

It also changes OhmPluginInfo to OhmPluginDesc, cutting down the functions a
plugin author needs to export to initialize, destroy and notify - of course,
all optional.

As the keys are all now const data, this cuts down strcpys dramitically so
memory use and fragmentation should be better. This also fixes a few memory
leaks.