add initializing stage to ohm-conf
Adds an initializing state to OhmConf in which no key-changed signals are
emitted. Any keys changed in this state are flagged as touched. When switched
out of initialising state, key-changed signals are emitted for all touched
keys.

Modifies OhmModule to switch the OhmConf into initalising state when calling
the init method of the loaded modules, swithing back to normal when all modules
have been initialised.