Added Nokia policy proto plugins (needs out prototype policy engine).
The plugins would need a fair amount of cleanup.
The centralized policy plugin is now the only one linked against the
set- and relation-manipulation libraries. It then exports those to
other plugins via out ohm-level method provides/requires-mechanism.
This is not needed (we could just link all the other plugins against
the shared libs) but it was a good test case for our ohm additions.