Preliminary support for runtime-configurable debugging/tracing support.
The current implementation is built as a set of macros on top of our
old NSN tracing library, libtrace. Some of the features available in
libtrace are probably an overkill or simply not needed (eg. multiple
trace contexts, trace tags, tag-based filtering). The provided macros
aim to hide the unnecessary features and the associated complexity.
functionality should be provided by the core OHM infrastructure. The
macros also aim to protect us from directly depending on libtrace making
easier to switch should OHM provide an official infrastructure for plugin
debugging.