Added forgotten fedora/ohm.spec.in.
I also managed to find the problem that was bugging the heck out of
me for quite a while now. Packaging seemed to work okay under debian
but on RPM-based distros %__tmpdir was creeping into the binaries which
got caught by the check-buildroot post-build script on recent distros.

It turns out the trivial obviousness of the culprit made it extremely
hard to spot although it was right there sticking into my eyes. I
accidentially omitted make from the %build section hence letting make
build everything from the %install section (%makeinstall). This caused
everything to be built as a dependency (as opposed to just getting
make installed) relocated under the staging directory. Duh... The same
problem exists (or existed) in all our other packages.