Support for plugin autodiscovery.
Added very primitive plugin autodiscovery and loading. If
ModulesRequired is set to "*" then the plugin directory is
searched for regular files with names mathing libohm_*.so.
The ones not in the banned list are added to ModulesRequired.
This is not exactly having acceptably modular configuration
but we get rid of our plugin packaging vs. monolithic modules.ini
problems without ugly/hairy postinstall kludgery.