Added support for specifying (plugin) trace flags on the command line.
Unfortunately it is still not possible to get init-time traces from
the plugins with this mechanism. The problem is that trace flags are
registered by the plugins themselves during plugin initialization which
is done in a single phase: plugins are dlopened and initialized in
dependency-order. Hence the trace flags for a plugin are not known
before the plugin has initialized itself.

One solution would be to improve a bit OHM plugin loading, and have
eg. something like

  1) load (dlopen) plugins
  2) resolve static cross-plugin methods, call plugin for minimal
     initialization (register trace flags, etc.)
  (at this point activate trace flags)
  3) call plugins for configuring themselves
  4) enter main loop