Honour implicit plugin dependencies stemming from importing plugin methods.
We do not perform a full topological sort based on both explicit and method-
based dependencies. Insted we start with a plugin list sorted by explicit
dependencies and resorder it by moving exporters before importers (of the same
method). Since we do not use (or need) explicit dependencies (ie.
OHM_PLUGIN_REQUIRES) much we do not run into conflicts and this works for us.
The pedantically correct thing would be to treat explicit and implicit
dependencies equally and sort plugins topologically considering both...