Plugin descriptors cannot be const (ie. potentially write-protected).
Since we change descriptors of plugins that import/export methods
from/to other plugins the descriptor cannot not be read-only. With
recent gcc/binutils/linker/loader in certain distributions (Ubuntu
for instance) symbols exported as const will always be mapped to
read-only sections of the address-space causing an immediate crash
during plugin loading.