Fix unit tests to compile with newer check library.
It's incorrect to assume that START_TEST directive creates a function
with a name that is exactly the same as a test name.
It was the case for older versions of check library, but this changed
in newer versions.

Relying on function names created by START_TEST directive is
relying on implementation detail that is undocumented and can change
at any time.

Hence, calling functions created by START_TEST directive from other tests
is incorrect.

This change moves all code from test functions that are supposed
to be called from other tests to static functions that are called
by these tests.

Signed-off-by: Oleg Girko <ol@infoserver.lv>